Support `@binaryAPI` on `private[this]` definitions
In this case the annotation will generate a public accessor (and setter).
This accessor has a consistent name that avoids clashes (`<className>$inline$<definitionName>`).

It will have the same name as a private accessor that would have been generate
in a non final class by the old inline accessor generation.