Generate kind-correct wildcards when selecting from a wildcard
Previously, derivedSelect would incorrectly approximate `fn.R` by `? >: Nothing
<: Any` in `i16997.min.scala`, which subsequently lead to a kind-incorrect
type application `Any[Y]` that broke type inference. #16999 proposed to fix this
by changing type inference, but it seems like we don't need to do that if we
make sure our wildcards are always kind-correct.

Note that the wildcard logic in lookupRefined could be removed: it seems like
it was supposed to be used by `reduceProjection` according to its documentation,
but the constructor of `NamedType` disallows wildcards as prefixes so it didn't
actually do anything there.

Co-Authored-By: Dale Wijnand <dale.wijnand@gmail.com>