Use "explain=" to tag an -explain message in an implicitNotFound
This approach mirrors the format used in `@noWarn`. There it is possible
to tag encode the category and message in the string.
```
@nowarn("cat=deprecation&msg=Implementation ...")
```

The current implementation for `explain=` only allows this tag at the
start of the string. Making the message either a old-style message or an
message in the explain.

We could extend this to support `msg=` to be able to have both a custom
message and an explain. Furthermore we could insert some flags to
explicitly state how to display the original message (for example:
not-found=hidden, not-found=in-message, not-found=in-explain).