Avoid creation of `@SplicedType` quote local refrences
The type declarations annotated with `@SplicedType` are only meant for
types that come from outside the quote. If the reference to the `Type[T]`
is to a definition within the quote (i.e. its level is grater than 0) we
can use it directly. The `@SplicedType` for that type will be generated
in a future compilation phase when the `Type[T]` definition reaches
level 0.

Fixes #17026