Optimize captured type in splice hole
If we have a quote that defined a type `T` and use it in the contents of
the splice we want to pass a `Type[T]` to the hole.

This improves a case where we had a `List[T]` and the hole could end up
with a `Type[T]` and `Type[List[T]]`. We should only pass the `Type[T]`.