Implement sealed type variables
Abolish restriction that box/unbox cannot be done with `cap`.

Instead, do a deep check for `cap`

 - in type arguments for `sealed` type variables
 - in the types of mutable vars
 - in the type of `try` under the `saferExceptions` language import

The caps.unsafe.{unsafeBox, unsafeUnbox, unsafeBoxFunArg} methods are not longer
needed and are deprecated. Instead there is an annotation annotation.unchecked.uncheckedCaptures that
can be added to a mutable variable to turn off checking its type.

These changes in behavior are introduced in 3.3. Older source versions
still support the old behavior (which corresponds to the paper).

So to run examples in the paper as described there, they need a `-source 3.2` or a
language import.

    import language.`3.2`.