Spec: Rewrite the type system fundamentals.
There are two fundamental differences between the type sytems of
Scala 2 and 3, which made a fundamental rewrite desirable:

- Scala 2 defines path-dependent types in terms of type projections,
  whereas Scala 3 is path-dependent first, and has not general
  type projects.
- Scala 2 has ad hod higher-kinded type parameters and polymorphic
  type aliases, whereas Scala 3 has first-class type lambdas and
  is fundamentally higher order.

We therefore rewrite the entire type system section of the spec
to be higher-kinded and path-dependent first. In the process, we
also revise refined types to better fit the path-dependent nature.

I (sjrd) took the liberty to add myself as a contributor in the
index page of the spec, as I believe this constitutes a fundamental
contribution to the language specification.