docs: remove legacy information from core data structures (#17464)
This PR is taken from the conversation in
https://github.com/lampepfl/dotty/issues/4015.

> I had a look at it, I think we should just drop core-data-structures.
> There's nothing that's still valuable to know today.

So I dropped most of this, but I also kept a bit of background
information that I found interesting and moved it into the dotc-scalac
page where we have some more info about symbols, denotations, and
differences between dotc/scalac. I thought it fit in nicely there, but
let me know if you'd think it should go somewhere else.

closes #4015

~Based on the comment in
https://github.com/lampepfl/dotty/pull/16776#issuecomment-1542404541, I
made this against language reference stable. I'm assuming this is
correct?~