Fix pretty-printing of renamed lambda params
Previously, the expression

    val a: (x: Int) => Bar[x.type] = ??? : ((x: Int) => Foo[x.type])

lead to the following error:

    Found:    (x: Int) => Foo[x.type]
    Required: (x: Int) => Bar[x².type]

    where:    x  is a reference to a value parameter
              x² is a reference to a value parameter

There's two problems here:
1. In the second lambda, x is supposed to be renamed to x², but the binder
   wasn't renamed.
2. Since the binders introducing x are part of the printed types, there isn't
   any ambiguity and no renaming is actually needed.

(1.) is a straight-up bug in the printing logic that we fix. (2.) is maybe a
matter of opinion, but for clarity we choose to address it by using the same
superscript for every parameter reference with the same name whose binder has
been printed. This does meant that `(x: Int) => (x: Int) => x.type` won't be
disambiguated, but something like that would be better addressed by printing a
warning about name shadowing anyway.