docs: add in a note about elidable not being supported in Scala 3
This is a follow-up of the comment in
https://github.com/lampepfl/dotty/issues/15766#issuecomment-1492846839
to try and make it clear to Scala 3 users that while this API exists and
is visible in Scala 3, it's not supported by the actual compiler.
Therefore it's recommended to instead use the `inline if` alternative.
This just adds a note to the Scaladoc to that effect.

refs: https://github.com/lampepfl/dotty/issues/15766