refactor(docs): update the contributing docs
This makes a bunch of small updates the contributing docs that were
recently migrated over from the scala-lang site. I think there are
actually some more updates to be done, but much of these changes are
just structural. To sort of summarize this does the following:

- gets rid of old irrelevant things like the `checklist.sh`
- removes some instructions that are no longer relevant
- re-organizes the side bar to better group things by topic
- gets rid of redundant information and tries to group things better
- moves the scaladoc docs into the contributing docs

I'll add some more comments inline to further explain.

[skip community_build]