Fix implicit type used for conversions, as seen from usage
Using the minimisation of the specs2 failure that occurred in CI, the
conversion desired is `Conversion[String, Bar.this.Data]`.  But if we
use the given's symbol info, we end up with the type `Conversion[String,
Foo.this.Data]` which the generated tree will fail to adapt to.

Using the widening of the implicit candidate's TermRef will yield a
method type, as seen from the right prefix - `Bar.this.Data`.