Omit more prefixes in non-package module printing
When we pretty-print a type in a module class, we end up printing the
module class's symbol.  For instance in a reduction failure of
Tuple.Union, we end up printing Tuple.Fold, with Tuple being an internal
(ThisType) reference to the Tuple module class.  I tweaked the logic in
fullNameString so that we omit more prefixes, if the module is
non-package (and make it consistent across -Ytest-pickler).  The package
part is important, because we want to continue to get "import
scala.concurrent.duration..." instead of "import concurrent.duration..".

[Cherry-picked fbb08f743a21cd3acfc526dd7ed05cd298ea4615]