Fix guarding & converting SAM result type to a function type
In decomposeProtoFunction I wanted to reuse WithFunctionType by running
toFunctionType on the whole SAM and then dropping the first parameters.
But with result dependent functions that results in a refined type,
which is composed of:
  * the function with a result that is a non-depedent approximation
  * the original, non-function, method type

So I went back to calling toFunctionType on the method result type
specifically, like it used to be.

Which means I stopped using the SAMType.WithFunctionType extract, so I
merged it with its remaining usage with subtype checking and called the
SAMType.isSamCompatible.