chore(ci): add in cancellations for concurrent builds of the same branch
This PR introduces a change to the CI that we use in Metals that has
really come in handy. The problem that this attempts to address is that
sometimes during the day the Dotty CI gets really backed up. Many times
when you look at why this is, it's because sometimes people push their
branch 2 or 3 times shortly after one another will small changes. This
then adds 3 full runs to CI when the vast majority of the time, the
latest one is the one they actually care about. What this change does is
cancels the previous one to run the latest. This _only_ does this on
PRs, not on the main branch.