Widen scrutinees in TypeComparer's compareMatch
If the skolem-widened scrutinees are the same and the cases correspond,
then the match types are the same.

Also, with inlining the type embeded in inlining's Typed expression will
be without going through type-avoidance, so they won't be skolems.  So
we look to widen inline proxy term refs, in addition to skolems.

Co-Authored-By: Matt Bovel <matthieu@bovel.net>