Fix signature computation involving nested type variables
During typechecking, we sometimes need to compute the signature of a type
involving uninstantiated type variables or wildcards (in particular, because
signature matching is doen as part of subtype checking for methods), this is
handled with `tpnme.Uninstantiated` which is handled specially in `Signature`.

Before this commit, `sigName` only checked for wildcards and type variables
at the top-level of the type, even though nested types can still have
an impact on type erasure, in particular when they appear as part of:
- an intersection
- a union
- an underlying type of a derived value class
- the element type of an array type
- an element type of a tuple (... *: X *: ...)

We keep track of all these situations by returning `null` in
`TypeErasure#apply` when encountering a wildcard or uninstantiated type
variable, which we then propage upwards to the `sigName` call. This propagation
only happens for `sigName` calls (where `inSigName` is true), otherwise we
throw an assertion since erasure shouldn't normally be computed for
underdefined types.

[Cherry-picked c21133f1c5b39054031aed2f28a82f1e396a3f2d]