Move test
Pickling tests only work if the program is consistently compiled
with capture checking on or off. A language import alone will not
do it, since that will be visible when pickling but not when unpickling.

The reason is that pickler test does not scan for language imports, i.e.
it does not call CompilationUnit.force.