Fix StaleSymbol for path dependent result type in macro context
The reason for the stale symbols was the fact that they were derived
from the previous (suspended) run (by calling a method from that run),
and path depended types seem to use an optimized path utilizing initial
designator, unlike the other kinds of types, which seem to either not
need calculating denotation or the denotation there is able to update
it's own validity.

Since that is impossible to do for those path dependent types, if they
are invalid in the run recaluclate the Symbol, otherwise use
`currentSymbol` as before.

[Cherry-picked f3c2b1df6a1e7de896ea7be3461032cc3387024d]