Fix: Style for wiki syntax
The goal of this PR is to correct the style for wiki syntax when the -comment-syntax:wiki is activate.
I have 2 possibilities:
First-Change the tag(Done on this PR): So when the correct syntax is called I change the tag to strong,em,etc.. instead of using a span with a class.
Second-Change the CSS: We keep the span with a class and we add the correct CSS associated to the class.

[Cherry-picked ebd164c74d5e3ffdc874e0b7b419818dc9571e69]