Align ArrayBuilder resizing with ArrayBuffer
If the requested size is larger than double the current array lenght,
use the requested size.

Set `VM_MaxArraySize` to `Int.MaxValue - 8`, which seems to be
a common default. Use that in ArrayBuffer and ArrayBuilder.

To check if a size of more than Int.MaxValue is requested, check
for an Int overflow instead of using a Long.