Improve logic when to emit pattern type error
Improve logic when to emit "pattern type is incompatible with expected type"
error.

Fixes #18083

The whole thing is not very satisfactory. We have an approximation which has both
false positives and false negatives.

False positives: We are too lenient for numeric types and and/or types
False negatives: We ignore user-generated equals methods

The new approximation seems to be somewhat better than the old, but it's still an
approximation. It begs the question why we attempt to do this at all.