Allow to beta reduce curried function applications in quotes reflect
Previously, the curried functions with multiple applications
were not able to be beta-reduced in any way, which was unexpected.
Now we allow reducing any number of top-level function applications
for a curried function. This was also made clearer in the documentation
for the affected (Expr.betaReduce and Term.betaReduce) methods.

[Cherry-picked 447cdf48ef7f2cae11118ad9b31d7e62c2da3e43]