Handle capture checking language imports correctly for -Ytest-pickler
The pickler test did not track whether capture checking was on. This meant that files with

  import language.experimental.captureChecking

has to be compiled additionally with the capture checking setting

  -language.experimental.captureChecking

to pass pickler tests. This made it impossible to test mixed code bases where
some files require capture checking and others don't.

We now use the previous setting of `CompilationUnit#needsCaptureChecking` when unpickling.