Keep track which classes were compiled with capture checking
Similar to the treatment of WithPureFuns, we add an annotation WithCaptureChecks
to toplevel classes if they were compiled with capture checking on. We also
keep a flag on all classes so that we can find out fast whether the class was
capture checked or not. However, that flag is not pickled. It is instead
reconstituted by the unpickler when it sees the WtihCaptureChecks annotation.