Better interop with non-capture checked files
When interacting with files that are not captue checked, we should assume
that types can have arbitrary capture sets. This is analogous to interacting
with non-null checked files, where we also assume that types in the interface
can be null or not.

This commit enables interop for normal rechecking by decorating types coming
from sources that are not capture-checked with a special `Fluid` capture set.
This set both super- and subcaptures any other capture set.