Move stdlib tests to pos
That way we can compile mixed capture-checked and non-capture checked sources.
Also, add collection/Iterable.scala to test.

We had to add Map as well since there is an incompatibility between the CB sources
(from which we take the tests) and the build binaries (against which they are compiled)
there.