don't disable all scalacOptions, only selectively Ysafe-init and Xcheck-macros
In the community build runner, comments for the shapeless and protoquill
projects say to disable the Ysafe-init and Xcheck-macros options, but
the implementation disables all scalacOptions. This change makes the
implementation match the comments.

Also, the comment for protoquill incorrectly stated that Ysafe-init
needs to be disabled due to a buggy macro. The project works fine with
Ysafe-init, but the buggy macro requires Xcheck-macros disabled.

[Cherry-picked b94f7ec61bc57ce745c84dea82506a0fa14ed355]