Always assert YnoDeepSubtypes in monitoredIsSubType
Not doing so risks making the first assertion going unnoticed.  For
instance, tryToShow, which is called by our string interpolators, will
suppress "non-fatal" exception, like this assertion.  Which means that
it can hide the failure and show a change in behaviour between the first
and the subsequent calls to subtyping.