Fix how an inferred type var affects match analysis
This patch is to fix i17118.

The issue is that, during typing, the type var `?A` in `case G[?A](i)`,
originally refers to the parameter `A`, which is bound by `Foo`.  Then
it is instantiated to `X`, which isn't bound.  So then match analysis
thinks that the case only covers `G[X]` values, leaving `G[Foo]` values
unconvered.

The solution involves instead of instantiating `?A` to `X`,
instantiating it to a fresh pattern bound symbol (`A$1`), with bounds
`>: X <: Foo & X`, which preserves the needed upper-bound.

The other changes are in order to make that work, in particular make
pos/creative-gadt-constraints5a (which is a part of
neg/creative-gadt-constraints) compile without blowing up isSubType.