Make sure in CI that we do not unexpectedly fall back on legacy match type reduction.
We introduce a new flag `-Yno-legacy-match-types`, which forbids
the reduction of "legacy" match types. Like `-Yno-deep-subtypes`,
it is meant to be used in our CI. With it, we check that we do not
unexpectedly fall back on legacy match types in tests for which
the specced match types are enough.

Later, we should consider integrating that behavior with the source
level so that it reaches users.