Fix implicitArg position of for-comprehension methods pointing to yield/do
Fixed by first detecting if a tree was desugared from a
for-comprehension, and if so, setting the position of the context
parameter to the right method symbol in the for comprehension tree.

The detection mechanism was based off of a similar implementation made
for semanticdb generation (`isForSynthetic`), which seemed to be sound.