Fix #18484: Query every legal child of sealed class on children call
Sometimes, a macro or Mirror can be called/generated for a type that
may have not had its children registered by natural typechecking
procedures of the compiler. This meant that, in those cases, the Mirror
could not have been generated correctly, and macro could not obtain the
children of the class.

This was partially fixed in a previous PR by introducing a separate
procedure for querying children, outside of the regular typechecking
procedure. However, they was only queried in the owner symbol or in the
companion class, when non-enum sealed classes specifically, can
appear anywhere in the file. This commit aims to rectify this by
searching for any class that appears in the same file as the sealed
type.