Support completions for extension definition parameter (#18331)
Extension methods are extended into normal definitions.
Because of that typed trees don't include any information about the
extension method definition parameter:
```scala
extension (x: In@@)
```
In order to add completions, we check if there is an exact path to the
untyped tree, and if not, we fall back to it. There may also be more
possible cases like that, but I can't think of any at the moment.
[Cherry-picked de4ad2b6c0bef31ee4a3750572222e95fcd9a278]