Handle outer class roots when instantiating class members
The problem arises if we have a class like the one in pos-custom-args/captures/refs.scala:

```scala
class MonoRef(init: Proc):
  type MonoProc = Proc
  var x: MonoProc = init
  def getX: MonoProc = x
  def setX(x: MonoProc): Unit = this.x = x
```

The type of `getX` and `setX` refer to the local root capability of class `MonoRef`.
When we call `m.getX` or `m.setX` in `m: MonoRef`, these occurrences have to be adapted to
capture roots in the scope of the selection. We determine these roots by inspecting the
capture set of `m` and picking a root that corresponds to it.