Keep track of enclosing capture sets when comparing refined types
When comparing refined types

   C{x: T}^cs  <:<  C{x: T'}^cs'

we need to remember the capture set `cs` on the left hand side when we
get to ask whether the refined type has a member `x: T'`, so that we
can correctly instantiate any local root `cap[C]` in the member type of `x`.
Same if the LHS is simply `C^cs`. In both cases we strip the capture set before
we get to take the compute `x`, so we have to re-add it at the point of that
member computation.