Drop unnecessary code in transformSym
The only time a nested object class could have a self type with a non-empty
capture set is when it declared in the current compilation unit. Then that
capture set is a Var that gets installed via an updateInfo. But then we do
also an updateInfo of the module val, so no need to do the same thing in
transformSym.